requirements.txt
asgiref==3.7.2
attrs==23.2.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.38.0
contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
deepmerge==1.1.1
Deprecated==1.2.14
dj-database-url==2.1.0
Django==4.2.7
django-appconf==1.0.6
django-compressor==4.4
django-environ==0.11.2
django-paypal==2.0
dotmap==1.3.30
firebase-admin==6.4.0
fonttools==4.47.0
gcloud==0.18.3
google-api-core==2.15.0
google-api-python-client==2.115.0
google-auth==2.27.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.14.0
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.0
grpcio-status==1.60.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.6
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jwcrypto==1.5.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.8.2
mdurl==0.1.2
msgpack==1.0.7
numpy==1.26.3
oauth2client==4.1.3
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.4
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.20.0
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
Pyrebase4==4.7.1
python-dateutil==2.8.2
python-jwt==4.1.0
pytz==2023.3.post1
pyzbar==0.1.9
rcssmin==1.1.1
referencing==0.32.1
requests==2.29.0
requests-toolbelt==0.10.1
rich==13.7.0
rjsmin==1.2.1
rpds-py==0.16.2
rsa==4.9
screeninfo==0.8.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==1.26.18
whitenoise==6.6.0
wrapt==1.16.0
